Move search, filter and part of categories logic to Products
