+++ src/App.js
@@ -8,16 +8,9 @@ import Recipes from './components/Recipes/Recipes'
 import useCategories from './hooks/useCategories'
 import useProduct from './hooks/useProduct'
 import useRating from './hooks/useRating'
-import useSearch from './hooks/useSearch'
 
 function App({ initialProductData, initialRecipeData }) {
-  const {
-    filterCategories,
-    activeCategory,
-    setActiveCategory,
-    handleCategoryClick,
-    handleAddCategories,
-  } = useCategories()
+  const { handleAddCategories } = useCategories()
 
   const { ratings, handleAddRating } = useRating()
 
@@ -26,35 +19,14 @@ function App({ initialProductData, initialRecipeData }) {
     handleAddRating,
   })
 
-  const { searchQuery, handleSearchInput, searchResults } = useSearch({
-    productData,
-  })
-
-  let shownData
-  if (activeCategory !== '') {
-    shownData = productData.filter(product =>
-      product.categories.includes(activeCategory)
-    )
-  } else if (searchResults !== productData) {
-    shownData = searchResults
-  } else {
-    shownData = productData
-  }
-
   return (
     <Router>
       <Switch>
         <Route exact path="/">
           <Products
-            shownData={shownData}
-            onCategoryClick={handleCategoryClick}
-            activeCategory={activeCategory}
-            setActiveCategory={setActiveCategory}
-            filterCategories={filterCategories}
             onAddRating={handleAddRating}
             ratings={ratings}
-            searchQuery={searchQuery}
-            onSearchInput={handleSearchInput}
+            productData={productData}
           />
         </Route>
         <Route exact path="/product-form">
