+++ src/components/Search/Search.js
@@ -1,32 +1,8 @@
-import { useHistory } from 'react-router'
 import styled from 'styled-components/macro'
-import debounce from 'lodash.debounce'
-import { useCallback, useEffect } from 'react'
 
-function Search({ searchQuery, onSearchInput }) {
-  const history = useHistory()
-
-  function handleSubmit(event) {
-    history.push(`?search=${searchQuery}`)
-    event.preventDefault()
-  }
-
-  function updateUrl() {
-    history.push(`?search=${searchQuery}`)
-  }
-
-  // eslint-disable-next-line
-  const delayedQuery = useCallback(debounce(updateUrl, 1000), [searchQuery])
-
-  useEffect(() => {
-    delayedQuery()
-
-    // Cancel the debounce on useEffect cleanup.
-    return delayedQuery.cancel
-  }, [searchQuery, delayedQuery])
-
+function Search({ searchQuery, onSearchInput, onSubmit }) {
   return (
-    <form autoComplete="off" onSubmit={handleSubmit}>
+    <form autoComplete="off" onSubmit={onSubmit}>
       <label>
         <ScreenReaderOnly>Search bar</ScreenReaderOnly>
         <input
